Rename gqs, gms from thing-service to thing
