+++ nomad/grapl-core.nomad
@@ -732,26 +732,32 @@ job "grapl-core" {
   }
 
   // todo: move to a new graph-db.nomad
-  group "graph-query-service" {
+  group "graph-query" {
     network {
       mode = "bridge"
       dns {
         servers = local.dns_servers
       }
-      port "graph-query-service-port" {
+      port "graph-query-port" {
       }
     }
 
-    task "graph-query-service" {
+    task "graph-query" {
       driver = "docker"
 
       config {
-        image = var.container_images["graph-query-service"]
-        ports = ["graph-query-service-port"]
+        image = var.container_images["graph-query"]
+        ports = ["graph-query-port"]
+      }
+
+      template {
+        data        = var.observability_env_vars
+        destination = "observability.env"
+        env         = true
       }
 
       env {
-        GRAPH_QUERY_SERVICE_BIND_ADDRESS = "0.0.0.0:${NOMAD_PORT_graph-query-service-port}"
+        GRAPH_QUERY_SERVICE_BIND_ADDRESS = "0.0.0.0:${NOMAD_PORT_graph-query-port}"
         RUST_BACKTRACE                   = local.rust_backtrace
         RUST_LOG                         = var.rust_log
         GRAPH_DB_ADDRESSES               = var.graph_db.addresses
@@ -761,15 +767,15 @@ job "grapl-core" {
     }
 
     service {
-      name = "graph-query-service"
-      port = "graph-query-service-port"
+      name = "graph-query"
+      port = "graph-query-port"
       connect {
         sidecar_service {}
       }
 
       check {
         type     = "grpc"
-        port     = "graph-query-service-port"
+        port     = "graph-query-port"
         interval = "10s"
         timeout  = "3s"
       }
@@ -777,26 +783,32 @@ job "grapl-core" {
   }
 
   // todo: move to a new graph-db.nomad
-  group "graph-mutation-service" {
+  group "graph-mutation" {
     network {
       mode = "bridge"
       dns {
         servers = local.dns_servers
       }
-      port "graph-mutation-service-port" {
+      port "graph-mutation-port" {
       }
     }
 
-    task "graph-mutation-service" {
+    task "graph-mutation" {
       driver = "docker"
 
       config {
-        image = var.container_images["graph-mutation-service"]
-        ports = ["graph-mutation-service-port"]
+        image = var.container_images["graph-mutation"]
+        ports = ["graph-mutation-port"]
+      }
+
+      template {
+        data        = var.observability_env_vars
+        destination = "observability.env"
+        env         = true
       }
 
       env {
-        GRAPH_MUTATION_SERVICE_BIND_ADDRESS = "0.0.0.0:${NOMAD_PORT_graph-mutation-service-port}"
+        GRAPH_MUTATION_SERVICE_BIND_ADDRESS = "0.0.0.0:${NOMAD_PORT_graph-mutation-port}"
         RUST_BACKTRACE                      = local.rust_backtrace
         RUST_LOG                            = var.rust_log
         GRAPH_DB_ADDRESSES                  = var.graph_db.addresses
@@ -804,33 +816,39 @@ job "grapl-core" {
         GRAPH_DB_AUTH_USERNAME              = var.graph_db.username
 
         # upstreams
-        GRAPH_SCHEMA_MANAGER_CLIENT_ADDRESS = "http://${NOMAD_UPSTREAM_ADDR_schema-manager}"
+        GRAPH_SCHEMA_MANAGER_CLIENT_ADDRESS = "http://${NOMAD_UPSTREAM_ADDR_graph-schema-manager}"
         UID_ALLOCATOR_ADDRESS               = "http://${NOMAD_UPSTREAM_ADDR_uid-allocator}"
       }
     }
 
     service {
-      name = "graph-mutation-service"
-      port = "graph-mutation-service-port"
+      name = "graph-mutation"
+      port = "graph-mutation-port"
       connect {
         sidecar_service {
           proxy {
             config {
               protocol = "grpc"
             }
-            # It'd be nice to dynamically use ports. Sadly, per https://github.com/hashicorp/nomad/issues/7135 its not
-            # to be. The ports chosen below can be changed at any time
+
             upstreams {
-              destination_name = "schema-manager"
-              local_bind_port  = 9999
+              destination_name = "graph-schema-manager"
+              local_bind_port  = 1000
             }
 
             upstreams {
               destination_name = "uid-allocator"
-              local_bind_port  = 9998
+              local_bind_port  = 1001
             }
           }
         }
+      }
+
+      check {
+        type     = "grpc"
+        port     = "graph-mutation-port"
+        interval = "10s"
+        timeout  = "3s"
       }
     }
   }
