+++ src/rust/graph-mutation/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "graph-mutation-service"
+name = "graph-mutation"
 version = "0.1.0"
 edition = "2021"
