+++ src/rust/graph-mutation/src/write_dropper.rs
