+++ src/rust/graph-query-service/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "graph-query-service"
+name = "graph-query"
 version = "0.1.0"
 edition = "2021"
